rename streampipes-processors-image-processing-jvm controllers to processors
