+++ streampipes-extensions/streampipes-processors-image-processing-jvm/src/main/java/org/apache/streampipes/processors/imageprocessing/jvm/processor/genericclassification/GenericImageClassificationProcessor.java
@@ -52,7 +52,7 @@ import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 
-public class GenericImageClassificationController extends StreamPipesDataProcessor {
+public class GenericImageClassificationProcessor extends StreamPipesDataProcessor {
 
   private String imagePropertyName;
