+++ streampipes-extensions/streampipes-processors-image-processing-jvm/src/main/java/org/apache/streampipes/processors/imageprocessing/jvm/processor/imageenrichment/ImageEnrichmentProcessor.java
@@ -15,8 +15,7 @@
  * limitations under the License.
  *
  */
-package org.apache.streampipes.processors.imageprocessing.jvm.processor.imagecropper;
-
+package org.apache.streampipes.processors.imageprocessing.jvm.processor.imageenrichment;
 
 import org.apache.streampipes.commons.exceptions.SpRuntimeException;
 import org.apache.streampipes.model.DataProcessorType;
@@ -25,7 +24,6 @@ import org.apache.streampipes.model.runtime.Event;
 import org.apache.streampipes.processors.imageprocessing.jvm.processor.commons.ImagePropertyConstants;
 import org.apache.streampipes.processors.imageprocessing.jvm.processor.commons.ImageTransformer;
 import org.apache.streampipes.processors.imageprocessing.jvm.processor.commons.RequiredBoxStream;
-import org.apache.streampipes.processors.imageprocessing.jvm.processor.imageenrichment.BoxCoordinates;
 import org.apache.streampipes.sdk.builder.ProcessingElementBuilder;
 import org.apache.streampipes.sdk.extractor.ProcessingElementParameterExtractor;
 import org.apache.streampipes.sdk.helpers.EpProperties;
@@ -38,34 +36,39 @@ import org.apache.streampipes.wrapper.routing.SpOutputCollector;
 import org.apache.streampipes.wrapper.standalone.ProcessorParams;
 import org.apache.streampipes.wrapper.standalone.StreamPipesDataProcessor;
 
+import java.awt.BasicStroke;
+import java.awt.Color;
+import java.awt.FontMetrics;
+import java.awt.Graphics2D;
+import java.awt.Rectangle;
+import java.awt.geom.Rectangle2D;
 import java.awt.image.BufferedImage;
 import java.util.Base64;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
-public class ImageCropperController extends StreamPipesDataProcessor {
-
+public class ImageEnrichmentProcessor extends StreamPipesDataProcessor {
   private String imageProperty;
   private String boxArray;
 
   @Override
   public DataProcessorDescription declareModel() {
-    return ProcessingElementBuilder.create("org.apache.streampipes.processor.imageclassification.jvm.image-cropper")
+    return ProcessingElementBuilder.create("org.apache.streampipes.processor.imageclassification.jvm.image-enricher")
         .withAssets(Assets.DOCUMENTATION, Assets.ICON)
         .withLocales(Locales.EN)
         .category(DataProcessorType.IMAGE_PROCESSING)
         .requiredStream(RequiredBoxStream.getBoxStream())
-        .outputStrategy(OutputStrategies.append(EpProperties.integerEp(Labels.empty(),
-                ImagePropertyConstants.CLASS_NAME.getProperty(), "https://streampipes.org/classname"),
-            EpProperties.doubleEp(Labels.empty(), ImagePropertyConstants.SCORE.getProperty(),
-                "https://streampipes.org/Label")))
+        .outputStrategy(OutputStrategies.fixed(
+            EpProperties.stringEp(Labels.empty(), ImagePropertyConstants.IMAGE.getProperty(),
+                "https://image.com")
+        ))
         .build();
   }
 
   @Override
   public void onInvocation(ProcessorParams parameters, SpOutputCollector spOutputCollector,
-                           EventProcessorRuntimeContext runtimeContext) {
+                           EventProcessorRuntimeContext runtimeContext) throws SpRuntimeException {
     ProcessingElementParameterExtractor extractor = parameters.extractor();
     this.imageProperty = extractor.mappingPropertyValue(RequiredBoxStream.IMAGE_PROPERTY);
     this.boxArray = extractor.mappingPropertyValue(RequiredBoxStream.BOX_ARRAY_PROPERTY);
@@ -78,33 +81,46 @@ public class ImageCropperController extends StreamPipesDataProcessor {
 
     if (imageOpt.isPresent()) {
       BufferedImage image = imageOpt.get();
-      List<Map<String, Object>> allBoxCoordinates = imageTransformer.getAllBoxCoordinates(boxArray);
+      List<Map<String, Object>> allBoxesMap = imageTransformer.getAllBoxCoordinates(boxArray);
 
-      for (Map<String, Object> box : allBoxCoordinates) {
-        BoxCoordinates boxCoordinates = imageTransformer.getBoxCoordinates(image, box);
+      for (Map<String, Object> box : allBoxesMap) {
+        BoxCoordinates boxCoordinates = imageTransformer.getBoxCoordinatesWithAnnotations(image, box);
 
-        BufferedImage dest = image.getSubimage(boxCoordinates.getX(), boxCoordinates.getY(), boxCoordinates.getWidth(),
-            boxCoordinates.getHeight());
+        Graphics2D graph = image.createGraphics();
 
-        Optional<byte[]> finalImage = imageTransformer.makeImage(dest);
+        // set color
+        Color color = ColorUtil.getColor(boxCoordinates.getClassesindex().hashCode());
+        graph.setColor(color);
 
-        if (finalImage.isPresent()) {
-          Event outEvent = new Event();
+        // Box
+        graph.setStroke(new BasicStroke(5));
+        graph.draw(new Rectangle(boxCoordinates.getX(), boxCoordinates.getY(), boxCoordinates.getWidth(),
+            boxCoordinates.getHeight()));
 
-          outEvent.addField(ImagePropertyConstants.TIMESTAMP.getProperty(),
-              in.getFieldByRuntimeName(ImagePropertyConstants.TIMESTAMP.getProperty()).getAsPrimitive().getAsLong());
+        // Label
+        String str = boxCoordinates.getClassesindex() + ": " + boxCoordinates.getScore();
 
-          outEvent.addField(ImagePropertyConstants.IMAGE.getProperty(),
-              Base64.getEncoder().encodeToString(finalImage.get()));
+        FontMetrics fm = graph.getFontMetrics();
+        Rectangle2D rect = fm.getStringBounds(str, graph);
 
-          outEvent.addField(ImagePropertyConstants.CLASS_NAME.getProperty(),
-              box.get(ImagePropertyConstants.CLASS_NAME.getProperty()));
+        graph.fillRect(boxCoordinates.getX(),
+            boxCoordinates.getY() - fm.getAscent(),
+            (int) rect.getWidth(),
+            (int) rect.getHeight());
 
-          outEvent.addField(ImagePropertyConstants.SCORE.getProperty(),
-              box.get(ImagePropertyConstants.SCORE.getProperty()));
+        graph.setColor(Color.white);
+        graph.drawString(str, boxCoordinates.getX(), boxCoordinates.getY());
 
-          out.collect(outEvent);
-        }
+        graph.dispose();
+      }
+
+      Optional<byte[]> finalImage = imageTransformer.makeImage(image);
+
+      if (finalImage.isPresent()) {
+        Event outEvent = new Event();
+        outEvent.addField(ImagePropertyConstants.IMAGE.getProperty(),
+            Base64.getEncoder().encodeToString(finalImage.get()));
+        out.collect(outEvent);
       }
     }
   }
